requirements.txt
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
tensorflow==2.13.0
nltk==3.8.1
matplotlib==3.7.2
seaborn==0.12.2
wordcloud==1.9.2
pyyaml==6.0.1
